feat(lapis): more suggestions
