+++ lapis/src/test/kotlin/org/genspectrum/lapis/DummyDatabaseConfig.kt
@@ -1,13 +1,13 @@
 package org.genspectrum.lapis
 
-import org.genspectrum.lapis.config.ADVANCED_QUERY_FIELD
 import org.genspectrum.lapis.config.DatabaseConfig
 import org.genspectrum.lapis.config.DatabaseFeature
 import org.genspectrum.lapis.config.DatabaseMetadata
 import org.genspectrum.lapis.config.DatabaseSchema
+import org.genspectrum.lapis.config.GENERALIZED_ADVANCED_QUERY_FEATURE
 import org.genspectrum.lapis.config.MetadataType
 import org.genspectrum.lapis.config.OpennessLevel
-import org.genspectrum.lapis.config.VARIANT_QUERY_FIELD
+import org.genspectrum.lapis.config.SARS_COV2_VARIANT_QUERY_FEATURE
 
 val dummyDatabaseConfig = DatabaseConfig(
     DatabaseSchema(
@@ -23,6 +23,9 @@ val dummyDatabaseConfig = DatabaseConfig(
             DatabaseMetadata("test_boolean_column", MetadataType.BOOLEAN),
         ),
         "some_metadata",
-        listOf(DatabaseFeature(VARIANT_QUERY_FIELD), DatabaseFeature(ADVANCED_QUERY_FIELD)),
+        listOf(
+            DatabaseFeature(name = SARS_COV2_VARIANT_QUERY_FEATURE),
+            DatabaseFeature(name = GENERALIZED_ADVANCED_QUERY_FEATURE),
+        ),
     ),
 )
